move opcodes to their own class to clear up clutter
